Move main loop to a separate file
